CURLOPT: replace `(long)` cast with `L` suffix for `CURLHSTS_*` macros
In curl/curl.h and the man page. To match the rest of CURLOPT macros.

Cherry-picked from #18130

+++ include/curl/curl.h
@@ -1067,8 +1067,8 @@ typedef CURLSTScode (*curl_hstswrite_callback)(CURL *easy,
                                                void *userp);
 
 /* CURLHSTS_* are bits for the CURLOPT_HSTS option */
-#define CURLHSTS_ENABLE       (long)(1<<0)
-#define CURLHSTS_READONLYFILE (long)(1<<1)
+#define CURLHSTS_ENABLE       (1L<<0)
+#define CURLHSTS_READONLYFILE (1L<<1)
 
 /* The CURLPROTO_ defines below are for the **deprecated** CURLOPT_*PROTOCOLS
    options. Do not use. */
